% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calculate the mean and standard deviation for all numerical columns in a dataframe}
\usage{
calc_stats(data)
}
\arguments{
\item{data}{A dataframe containing the data to summarize}
}
\value{
a dataframe
}
\description{
Calculate the mean and standard deviation for all numerical columns in a dataframe
}
\examples{
df <- data.frame(
a = rnorm(5),
b = rnorm(5))
calc_stats(df)
}
