% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create a box plot}
\usage{
create_boxplot(data, x, y)
}
\arguments{
\item{data}{A dataframe containing the data to plot.}

\item{x}{A variable in the dataframe that will be on the plot's x axis.}

\item{y}{A variable in the dataframe that will be on the plot's y axis.}
}
\value{
a boxplot
}
\description{
Create a box plot
}
\examples{
df <- data.frame(
a = rnorm(5),
b = rnorm(5))
create_boxplot(df,"a","b")
}
