% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_dir.R
\name{create_output_dir}
\alias{create_output_dir}
\title{Creates a directory at the given path if it does not already exist}
\usage{
create_output_dir(output_dir)
}
\arguments{
\item{output_dir}{The path to the output directory}
}
\description{
Creates a directory at the given path if it does not already exist
}
\examples{
create_output_dir('../results/')
}
