% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter.R
\name{create_scatter}
\alias{create_scatter}
\title{Create a scatter plot}
\usage{
create_scatter(data, scatter1, scatter2, color_var)
}
\arguments{
\item{data}{A dataframe containing the data to plot.}

\item{scatter1}{A variable in the dataframe that will be on the plot's x axis.}

\item{scatter2}{A variable in the dataframe that will be on the plot's y axis.}

\item{color_var}{A variable in the dataframe that will be used to color the points on the plot.}
}
\value{
a scatterplot
}
\description{
Create a scatter plot
}
\examples{
df <- data.frame(
a = rnorm(5),
b = rnorm(5),
c = rnorm(5))
create_scatter(df,"a","b","c")
}
