% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_workflow.R
\name{model_workflow}
\alias{model_workflow}
\title{Simplifies the workflow for fitting and training a knn model}
\usage{
model_workflow(train_data, num_folds, range_neighbors, response)
}
\arguments{
\item{train_data}{The training dataset}

\item{num_folds}{The desired number of folds to use for cross validation}

\item{range_neighbors}{A sequence of the desired values of k}

\item{response}{The response variable to use in the model's formula}
}
\value{
a fitted model
}
\description{
Simplifies the workflow for fitting and training a knn model
}
\examples{
set.seed(123)
train_data <- data.frame(
a = rnorm(100),
b = rnorm(100),
response = sample(c('a', 'b', 'c'), 100, replace = TRUE)
)
num_folds <- 5
range_neighbors <- seq(1, 20)
response <- "response"
model_workflow(train_data,num_folds, range_neighbors, response)
}
