% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_factor.R
\name{read_and_factor}
\alias{read_and_factor}
\title{Read data at a given directory and factor columns if needed}
\usage{
read_and_factor(input_dir, factor_vars = NULL)
}
\arguments{
\item{input_dir}{The directory where the file to be read is located}

\item{factor_vars}{a single string or vector of strings representing variables to factor}
}
\value{
a dataframe
}
\description{
Read data at a given directory and factor columns if needed
}
\examples{

}
