% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_workflow.R
\name{model_workflow}
\alias{model_workflow}
\title{Simplifies the workflow for fitting and training a knn model}
\usage{
model_workflow(data, num_folds, range_neighbors, response)
}
\arguments{
\item{data}{The dataset to be split into training and testing sets}

\item{num_folds}{The desired number of folds to use for cross validation}

\item{range_neighbors}{A sequence of the desired values of k}
}
\value{
a fitted model
}
\description{
Simplifies the workflow for fitting and training a knn model
}
\examples{
data <- data.frame(
a = rnorm(1000),
b = rnorm(1000),
cultivar = sample(c(1,2,3), 1000, replace = TRUE)
)
num_folds <- 5
range_neighbors <- seq(1, 20)
response <- "cultivar"
model_workflow(data, num_folds, range_neighbors, response)
}
